style(clippy): green the rust CI job — resolve 424 clippy findings across 124 crates
cargo clippy --workspace --all-targets -- -D warnings (CI's rust job, after
fmt) was red across the generated crate set, never linted. Fixed with clippy
0.1.88 (CI toolchain): two --fix passes + one --unsafe-fixes pass for the bulk
(collapsible_if, manual_*, unnecessary_*, doc_*), then manual residual:
- 11 intentional inherent methods (#[allow(should_implement_trait)] on widget
  next(); #[allow(too_many_arguments)] on a 10-arg + 8-arg constructor)
- ItemPin::is_empty() added; 3x get().is_none() -> contains_key; index loop ->
  slice iter; max().min() -> clamp(); 2 nested format! flattened; 2
  if-identical-blocks collapsed (behaviour-preserving, verified non-bugs)
- 10 rustdoc list-formatting lints fixed
- caught + fixed a clippy --fix over-reach: it dropped a cfg(test)-only
  Modifiers import from shortcut-cheatsheet; re-imported in the test module

clippy exits 0; cargo fmt --check clean. 126 files, all behaviour-preserving.

+++ crates/sovereign-cockpit-marquee-loop/src/lib.rs
@@ -92,7 +92,7 @@ impl MarqueeLoop {
         let cycle_px = text_px.saturating_add(self.gap_px);
         // offset = (speed * t_ms / 1000) mod cycle_px
         let micro_px = (self.speed_px_per_s as u128) * (t_ms as u128);
-        let px = (micro_px / 1000u128) as u128;
+        let px = micro_px / 1000u128;
         let x_offset_px = (px % (cycle_px as u128)) as u32;
         Frame {
             state: State::Looping,
